+++ scripts/preprocessing/bcdi_read_data_P10.py
@@ -25,7 +25,8 @@ from scipy.io import savemat
 import sys
 import bcdi.graph.graph_utils as gu
 import bcdi.utils.utilities as util
-import bcdi.experiment.experiment_utils as exp
+from bcdi.experiment.detector import Detector
+
 
 helptext = """
 Open images or series data at P10 beamline.
@@ -175,7 +176,7 @@ def main(parameters):
     #######################
     # Initialize detector #
     #######################
-    detector = exp.Detector(name=parameters["detector"])
+    detector = Detector(name=parameters["detector"])
     nb_pixel_y, nb_pixel_x = detector.nb_pixel_y, detector.nb_pixel_x
     sumdata = np.zeros((nb_pixel_y, nb_pixel_x))
     mask = np.zeros((nb_pixel_y, nb_pixel_x))
